requirements.txt
itsdangerous==0.24
gunicorn==20.0.4
jinja2==2.10
markupsafe==1.0
werkzeug==0.14.1
flask==1.0.2 	
matplotlib>=2.2.3 
requests>=2.19.1 
pandas>=0.23.4 
scikit-learn >=0.23.1                  
scipy>=1.1.0            
seaborn>=0.9.0        
joblib>=0.14.1   
numpy>=1.15.4    
nltk>=3.3.0  
tweet-preprocessor>=0.6.0       
cloudpickle>=0.5.5 
keras>=2.2.4 
keras-preprocessing>=1.1.0                     
numpy-base>=1.15.4           
numpydoc>=0.8.0                   
requests-oauthlib>=1.3.0
unicodecsv>=0.14.1                  
urllib3>=1.23                    
wordcloud>=1.7.0                  